Relicense to GPLv2
I originally chose the LGPL thinking that this mod would have an API that other
mods could use. That no longer appears to be the case, so I'm switching to the
GPL from this point. Note that since the LGPL is irrevocable, any commits made
before this one will continue to be made available under it; only future
changes will be GPL-only.